+++ libs/common/src/lib/interfaces/responses/public-portfolio-response.interface.ts
@@ -14,32 +14,10 @@ export interface PublicPortfolioResponse extends PublicPortfolioResponseV1 {
     [symbol: string]: Pick<
       PortfolioPosition,
       | 'allocationInPercentage'
-
-      /** @deprecated */
-      | 'assetClass'
       | 'assetProfile'
-
-      /** @deprecated */
-      | 'countries'
-      | 'currency'
-
-      /** @deprecated */
-      | 'dataSource'
       | 'dateOfFirstActivity'
       | 'markets'
-
-      /** @deprecated */
-      | 'name'
       | 'netPerformancePercentWithCurrencyEffect'
-
-      /** @deprecated */
-      | 'sectors'
-
-      /** @deprecated */
-      | 'symbol'
-
-      /** @deprecated */
-      | 'url'
       | 'valueInBaseCurrency'
       | 'valueInPercentage'
     >;